Disallow undefined arguments to update cmds (#191)

+++ packages/mapper/src/mapper.ts
@@ -492,6 +492,10 @@ export default class Mapper {
     const dbOptions: DatabaseOptions = {};
     const db = collection._database._name;
     const coll = collection._name;
+
+    if (filter === undefined) {
+      throw new MongoshInvalidInputError('Cannot pass an undefined argument to an update command');
+    }
     this.messageBus.emit(
       'mongosh:api-call',
       {
@@ -536,6 +540,9 @@ export default class Mapper {
     const dbOptions: DatabaseOptions = {};
     const db = collection._database._name;
     const coll = collection._name;
+    if (filter === undefined) {
+      throw new MongoshInvalidInputError('Cannot pass an undefined argument to an update command');
+    }
     this.messageBus.emit(
       'mongosh:api-call',
       {
@@ -792,6 +799,10 @@ export default class Mapper {
       }
     );
 
+    if (filter === undefined) {
+      throw new MongoshInvalidInputError('Cannot pass an undefined filter to an update command');
+    }
+
     const result = await this.serviceProvider.findOneAndDelete(
       db,
       coll,
@@ -825,6 +836,10 @@ export default class Mapper {
     if ('returnNewDocument' in findOneAndReplaceOptions) {
       findOneAndReplaceOptions.returnDocument = findOneAndReplaceOptions.returnNewDocument;
       delete findOneAndReplaceOptions.returnNewDocument;
+    }
+
+    if (filter === undefined || replacement === undefined) {
+      throw new MongoshInvalidInputError('Cannot pass an undefined argument to an update command');
     }
 
     this.messageBus.emit(
@@ -872,6 +887,10 @@ export default class Mapper {
       delete findOneAndUpdateOptions.returnNewDocument;
     }
 
+    if (filter === undefined || update === undefined) {
+      throw new MongoshInvalidInputError('Cannot pass an undefined argument to an update command');
+    }
+
     this.messageBus.emit(
       'mongosh:api-call',
       {
@@ -913,6 +932,9 @@ export default class Mapper {
 
     if ('writeConcern' in options) {
       Object.assign(dbOptions, options.writeConcern);
+    }
+    if (docs === undefined) {
+      throw new MongoshInvalidInputError('Cannot pass an undefined argument to an update command');
     }
 
     this.messageBus.emit(
@@ -957,6 +979,9 @@ export default class Mapper {
     const db = collection._database._name;
     const coll = collection._name;
 
+    if (docs === undefined) {
+      throw new MongoshInvalidInputError('Cannot pass an undefined argument to an update command');
+    }
     if ('writeConcern' in options) {
       Object.assign(dbOptions, options.writeConcern);
     }
@@ -1003,6 +1028,9 @@ export default class Mapper {
     const db = collection._database._name;
     const coll = collection._name;
 
+    if (doc === undefined) {
+      throw new MongoshInvalidInputError('Cannot pass an undefined argument to an update command');
+    }
     if ('writeConcern' in options) {
       Object.assign(dbOptions, options.writeConcern);
     }
@@ -1065,6 +1093,9 @@ export default class Mapper {
     const db = collection._database._name;
     const coll = collection._name;
 
+    if (query === undefined) {
+      throw new MongoshInvalidInputError('Cannot pass an undefined argument to an update command');
+    }
     if ('writeConcern' in options) {
       Object.assign(dbOptions, options.writeConcern);
     }
@@ -1099,6 +1130,9 @@ export default class Mapper {
     const dbOptions: DatabaseOptions = {};
     const db = collection._database._name;
     const coll = collection._name;
+    if (doc === undefined) {
+      throw new MongoshInvalidInputError('Cannot pass an undefined argument to an update command');
+    }
     this.messageBus.emit(
       'mongosh:api-call',
       {
@@ -1135,6 +1169,9 @@ export default class Mapper {
     const dbOptions: DatabaseOptions = {};
     const db = collection._database._name;
     const coll = collection._name;
+    if (filter === undefined) {
+      throw new MongoshInvalidInputError('Cannot pass an undefined argument to an update command');
+    }
     this.messageBus.emit(
       'mongosh:api-call',
       {
@@ -1168,6 +1205,9 @@ export default class Mapper {
   async collection_update(collection, filter, update, options: any = {}): Promise<any> {
     const db = collection._database._name;
     const coll = collection._name;
+    if (filter === undefined) {
+      throw new MongoshInvalidInputError('Cannot pass an undefined argument to an update command');
+    }
     this.messageBus.emit(
       'mongosh:api-call',
       {
@@ -1220,6 +1260,9 @@ export default class Mapper {
    * @returns {UpdateResult} The promise of the result.
    */
   async collection_updateMany(collection, filter, update, options: any = {}): Promise<any> {
+    if (filter === undefined) {
+      throw new MongoshInvalidInputError('Cannot pass an undefined argument to an update command');
+    }
     const dbOptions: DatabaseOptions = {};
     const db = collection._database._name;
     const coll = collection._name;
@@ -1273,6 +1316,9 @@ export default class Mapper {
     update: Document,
     options: Document = {}
   ): Promise<any> {
+    if (filter === undefined) {
+      throw new MongoshInvalidInputError('Cannot pass an undefined argument to an update command');
+    }
     this._emitCollectionApiCall(collection, 'updateOne', { filter, options });
 
     const dbOptions: DatabaseOptions = {};
@@ -1342,6 +1388,9 @@ export default class Mapper {
     keyPatterns: Document[],
     options: Document = {}
   ): Promise<any> {
+    if (keyPatterns === undefined) {
+      throw new MongoshInvalidInputError('Cannot pass an undefined argument to an update command');
+    }
     const db = collection._database._name;
     const coll = collection._name;
 
@@ -1382,6 +1431,9 @@ export default class Mapper {
     keys: Document,
     options: Document = {}
   ): Promise<any> {
+    if (keys === undefined) {
+      throw new MongoshInvalidInputError('Cannot pass an undefined argument to an update command');
+    }
     this.messageBus.emit(
       'mongosh:api-call',
       {
@@ -1540,6 +1592,9 @@ export default class Mapper {
     collection: Collection,
     indexes: string|string[]|Document|Document[]
   ): Promise<any> {
+    if (indexes === undefined) {
+      throw new MongoshInvalidInputError('Cannot pass an undefined argument to an update command');
+    }
     const db = collection._database._name;
     const coll = collection._name;
     this.messageBus.emit(
